revert addition of donate form in checkout
