+++ src/components/donation/Steps/Submit/StripeCheckout/Checkout.tsx
@@ -1,51 +1,93 @@
-import { Elements } from "@stripe/react-stripe-js";
-import { loadStripe } from "@stripe/stripe-js";
-import { PUBLIC_STRIPE_KEY } from "constants/env";
-import { useStripePaymentIntentQuery } from "services/apes";
-import { StripeCheckoutStep } from "slices/donation";
-import Err from "../Err";
+import {
+  PaymentElement,
+  useElements,
+  useStripe,
+} from "@stripe/react-stripe-js";
+import LoadText from "components/LoadText";
+import { GENERIC_ERROR_MESSAGE } from "constants/common";
+import { appRoutes, donateWidgetRoutes } from "constants/routes";
+import { useErrorContext } from "contexts/ErrorContext";
+import { FormEventHandler, useState } from "react";
+import { DonationSource } from "types/lists";
 import Loader from "../Loader";
-import CheckoutForm from "./CheckoutForm";
 
-type Donor = {
-  firstName: string;
-  lastName: string;
-  email: string;
-};
+// Code inspired by React Stripe.js docs, see:
+// https://stripe.com/docs/stripe-js/react#useelements-hook
+export default function Checkout({ source }: { source: DonationSource }) {
+  const stripe = useStripe();
+  const elements = useElements();
+  const { handleError } = useErrorContext();
 
-const stripePromise = loadStripe(PUBLIC_STRIPE_KEY);
+  const [isLoading, setLoading] = useState(true);
+  // There is a small delay before Stripe Payment Element starts to load.
+  // To avoid just showing the "Back" button with nothing else on screen,
+  // we first show a Loader ring and when the Stripe Element starts loading
+  // (it has an inherent loading animation) that's when we hide the loader ring
+  // and start showing the "Back" button
+  const [showLoader, setShowLoader] = useState(true);
+  const [isSubmitting, setSubmitting] = useState(false);
 
-// Followed Stripe's custom flow docs
-// https://stripe.com/docs/payments/quickstart
-export default function Checkout(props: StripeCheckoutStep & { donor: Donor }) {
-  const { details, recipient, liquidSplitPct } = props;
-  const {
-    data: clientSecret,
-    isLoading,
-    isError,
-    error,
-  } = useStripePaymentIntentQuery({
-    amount: +details.amount,
-    currency: details.currency.code,
-    endowmentId: recipient.id,
-    email: props.donor.email,
-    splitLiq: liquidSplitPct.toString(),
-    donorFullName: `${props.donor.firstName} ${props.donor.lastName}`,
-  });
+  const handleSubmit: FormEventHandler<HTMLFormElement> = async (e) => {
+    e.preventDefault();
 
-  return isLoading ? (
-    <Loader msg="Loading payment form.." />
-  ) : isError || !clientSecret ? (
-    <Err error={error} />
-  ) : (
-    <Elements
-      options={{
-        clientSecret,
-        appearance: { theme: "stripe" },
-      }}
-      stripe={stripePromise}
-    >
-      <CheckoutForm source={details.source} />
-    </Elements>
+    if (!stripe || !elements) {
+      // Stripe.js hasn't yet loaded.
+      // Make sure to disable form submission until Stripe.js has loaded.
+      return;
+    }
+
+    setSubmitting(true);
+
+    const return_url =
+      source === "bg-widget"
+        ? `${window.location.origin}${appRoutes.donate_widget}/${donateWidgetRoutes.stripe_payment_status}`
+        : `${window.location.origin}${appRoutes.stripe_payment_status}`;
+
+    const { error } = await stripe.confirmPayment({
+      elements,
+      confirmParams: {
+        return_url,
+      },
+    });
+
+    // This point will only be reached if there is an immediate error when
+    // confirming the payment. Otherwise, your customer will be redirected to
+    // your `return_url`. For some payment methods like iDEAL, your customer will
+    // be redirected to an intermediate site first to authorize the payment, then
+    // redirected to the `return_url`.
+    if (error.type === "card_error" || error.type === "validation_error") {
+      handleError(error.message);
+    } else {
+      handleError("An unexpected error occurred.");
+    }
+
+    setSubmitting(false);
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="contents">
+      <PaymentElement
+        options={{ layout: "tabs" }}
+        onReady={() => setLoading(false)}
+        onLoadError={(error) => {
+          setLoading(false);
+          handleError(error, GENERIC_ERROR_MESSAGE);
+        }}
+        onLoaderStart={() => setShowLoader(false)}
+      />
+      {showLoader ? (
+        <Loader />
+      ) : (
+        <button
+          className="btn-orange btn-donate w-full mt-6"
+          disabled={!stripe || !elements || isSubmitting || isLoading}
+          type="submit"
+        >
+          <LoadText text="Processing..." isLoading={isSubmitting}>
+            Pay Now
+          </LoadText>
+        </button>
+      )}
+    </form>
   );
 }
