+++ src/components/donation/Steps/Submit/Submit.tsx
@@ -24,7 +24,13 @@ export default function Submit(props: SubmitStep) {
 
   if (props.details.method === "stripe") {
     const { details, ...rest } = props;
-    return <StripeCheckout details={details} {...rest} />;
+    return (
+      <StripeCheckout
+        details={details}
+        {...rest}
+        donor={{ firstName: "", lastName: "", email: "" }}
+      />
+    );
   }
 
   if (props.details.method === "chariot") {
